Complete Discovery module data layer with working SQLX cache
- Fix priority type to i64 for SQLite compatibility
- Regenerate SQLX query cache for all discovery queries
- Add Discovery types to TypeScript generation
- Task type now includes discovery_item_id field

The data layer is now complete and compiles. Next steps:
- Add DiscoveryService business logic
- Add API routes for /api/discovery/*
- Add Discovery tab to frontend TaskPanel

+++ crates/db/src/models/discovery_item.rs
@@ -48,7 +48,7 @@ pub struct DiscoveryItem {
     pub acceptance_criteria: Option<String>,
     #[sqlx(json)]
     pub effort_estimate: Option<String>, // JSON string, parsed client-side
-    pub priority: Option<i32>,
+    pub priority: Option<i64>,
     pub promoted_task_id: Option<Uuid>,
     pub parent_id: Option<Uuid>,
     pub created_at: DateTime<Utc>,
@@ -63,7 +63,7 @@ pub struct CreateDiscoveryItem {
     pub content: Option<String>,
     pub acceptance_criteria: Option<String>,
     pub effort_estimate: Option<String>,
-    pub priority: Option<i32>,
+    pub priority: Option<i64>,
     pub parent_id: Option<Uuid>,
 }
 
@@ -75,7 +75,7 @@ pub struct UpdateDiscoveryItem {
     pub content: Option<String>,
     pub acceptance_criteria: Option<String>,
     pub effort_estimate: Option<String>,
-    pub priority: Option<i32>,
+    pub priority: Option<i64>,
     pub parent_id: Option<Uuid>,
 }
